Fix btwrite timer timeout for Angler (1/2)
1. For vnd_angler.txt:
Set PROC_BTWRITE_TIMER_TIMEOUT_MS to 0
to avoid upio start and stop timer when LMP
is enabled and BT_WAKE is asserted and
de-asserted. This time is useless and would
set btwrite to 0 by mistake even when host
is sending data to UART, and this would
de-assert BT_WAKE by mistake.

2. For hardware.c:
For chipset 4358 we need to set LPM_HC_IDLE_THRESHOLD
to 24, so we need to reduce the multiplier
to make timeout_ms a reasonable value. Without
modification, the timeout_ms would be 36000ms
which is too large.

3. For upio.c:
Need to set lmp_proc_cb.btwrite_active even though
PROC_BTWRITE_TIMER_TIMEOUT_MS is set to 0.
Otherwise, btwrite would be written very frequently.

Bug: 23248798
Change-Id: Ie55c346785679e51a771d5e46270c0858e6082ee