Fix hardware timing needed for BCM chip when reseting
The specification of BCM BT chips indicate timing to
respect between the operations powerOFF/powerON/first-command in
order to the current regulator works properly.

This patch make sleep the process the amount of
time needed for the current regulator to works properly.
If there is an error to sleep, an error is logged.

Change-Id: Ib6cd82d7f93291c387728c2a269c9195aa6e2ba7
Signed-off-by: Franck Lenormand <franckx.lenormand@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>