Update patch set 1
Patch Set 1:

> > Quick question: shouldn't the rfkill driver in the kernel be the
 > > one adhering the timing for the particular chip? This libbt is
 > > generic to all broadcom chips.
 > 
 > In fact, this timing is specifically for all BCM chips in order to
 > their current regulator has enough time to transition (according to
 > the datasheet of the chip which can't be disclose).

Understood, however this feels like something that belongs in the domain of the rfkill driver in the kernel instead of libbt in userspace (since that kernel driver is responsible for all the other settle delays, etc)

By moving it into the rfkill driver, anyone else who pokes around at the rfkill node outside of libbt gets the correct behavior for free as well.

Is there a reason this needs to be done in libbt?

Patch-set: 1